+++ src/index.ts
@@ -1,5 +1,7 @@
 export * from './compareUint8Arrays';
 export * from './concatUint8Arrays';
 export * from './constants';
+export * from './fromHex';
+export * from './toHex';
 export * from './uint64BE';
 export * from './writeUInt32BE';